Struct primordial::Offset

#[repr(transparent)]
pub struct Offset<T, U>(_, _);

An offset of a number of items of type T from a base

Note well that this is NOT stored in memory as the number of bytes, but rather the number of items.

One important additional feature is that offsets can be converted between underlying types so long as the conversion is lossless for the target CPU architecture. For example, Offset<u64> can be converted to Offset<usize> on 64-bit systems.

Implementations

impl<T, U> Offset<T, U>

pub const fn from_items(items: T) -> Self

Create an offset value from the number of items

pub fn items(self) -> T

Get the number of items

impl<T, U> Offset<T, U> where
Offset<usize, ()>: Into<Offset<T, ()>>,
T: Mul<T, Output = T>,

pub fn bytes(self) -> T

Get the number of bytes