Struct primordial::Register

#[repr(transparent)]
pub struct Register<T>(_);

A register

This type is intended to be used wherever raw access to a register value is required. The type itself is opaque, but it can be converted to usable types.

One important additional feature is that registers can be converted between underlying types so long as the conversion does not truncate. For example, Register<u64> can be converted to Register<usize> on 64-bit systems. Likewise, Register<usize> can be converted to and from a pointer.

Implementations

impl<T> Register<T>

pub unsafe fn into_slice<'a, U>(self, len: impl Into<usize>) -> &'a [U] where
Self: Into<*const U>,

Converts a register value to a slice

Safety

This function is unsafe because we are converting an integer to a pointer and then dereferencing it. The caller MUST ensure that the value of this register points to valid memory.

pub unsafe fn into_slice_mut<'a, U>(self, len: impl Into<usize>) -> &'a mut [U] where
Self: Into<*mut U>,

Converts a register value to a mutable slice

Safety

This function is unsafe because we are converting an integer to a pointer and then dereferencing it. The caller MUST ensure that the value of this register is valid memory.